requirements.txt
Flask
Flask-Limiter
Flask-SQLAlchemy
SQLAlchemy
argon2-cffi
Pillow
Flask-Login
requests
gunicorn
boto3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
python-dotenv